requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
requests>=2.31.0
google-generativeai>=0.4.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
sentence-transformers>=2.3.1
beautifulsoup4>=4.12.3
jinja2>=3.1.3
python-multipart>=0.0.9
